Add files via upload
This was an ongoing project throughout the 2023 MLB season. My goal was to track Ronald Acuna Jr's pace toward a season with 40 homeruns and stolen bases. Through this project I learned about pulling data from an extranal source with the baseballr package, I progressively improved the graphical representation of the information, and I attempted to communicate the findings through the project. This was also uploaded several times to the r/sabermetrics and r/baseball subreddit forums for feedback, thus giving me the opportunity to gather and incorporate external comments. Currently, the project remains unfinished due to a change in the API of the fangraphs baseball stat site rendering the baseballr call inoperable. I hope to find a workaround for this so that the finishing touches can be put on this project that was very enjoyable to work on.